requirements.txt
datasets
pandas
openai
ollama
faiss-cpu
numpy
tqdm
pyarrow
rank-bm25
tqdm
torch
ipykernel
sentence-transformers